+++ src/Apps/Artist/Routes/AuctionResults/index.tsx
@@ -1,7 +1,7 @@
 import { AuctionResults_artist } from "__generated__/AuctionResults_artist.graphql"
 import React from "react"
 import { createFragmentContainer, graphql } from "react-relay"
-import { AuctionResultsRefetchContainer as AuctionResults } from "./AuctionResultsRefetchContainer"
+import { ArtistAuctionResultsRefetchContainer as AuctionResults } from "./ArtistAuctionResults"
 
 export interface AuctionResultsRouteProps {
   artist: AuctionResults_artist
@@ -17,7 +17,7 @@ export const AuctionResultsRouteFragmentContainer = createFragmentContainer(
   AuctionResultsRoute,
   graphql`
     fragment AuctionResults_artist on Artist {
-      ...AuctionResultsRefetchContainer_artist
+      ...ArtistAuctionResults_artist
     }
   `
 )
